Move cronjob email l10n code into l10n folder
